Fix clippy lints
In `group_files_by_date` the clone is not needed.
`or_insert_with`is more efficient in the case where a value already exists, because the closure or function provided to it are only executed if needed, while or_insert always creates the value.

BTW keep up the good work. I really enjoyed those posts! :)